Change S3 init
